+++ src/Application/Resources/Queries/GetConflictingResources/GetConflictingResourcesQuery.cs
@@ -50,7 +50,7 @@ namespace ResourcePackerGUI.Application.Resources.Queries.GetConflictingResource
         /// <summary>
         /// An optional progress instance to keep track of the checking process.
         /// </summary>
-        public IProgress<int>? Progress { get; init; }
+        public IProgress<float>? Progress { get; init; }
 
         /// <summary>
         /// The interval in milliseconds for updating the progress instances when present.
