+++ src/Application/Resources/Queries/UpdateResourceDefinitions/UpdateResourceDefinitionsQueryHandler.cs
@@ -43,7 +43,7 @@ namespace ResourcePackerGUI.Application.Resources.Queries.UpdateResourceDefiniti
             var updated = 0;
             using (var progressTimer = new System.Timers.Timer(request.ProgressReportInterval))
             {
-                var percentage = 0;
+                var percentage = 0f;
 
                 // ReSharper disable once AccessToModifiedClosure
                 progressTimer.Elapsed += delegate { request.Progress!.Report(percentage); };
@@ -51,7 +51,7 @@ namespace ResourcePackerGUI.Application.Resources.Queries.UpdateResourceDefiniti
 
                 for (var i = 0; i < request.Resources.Count; i++)
                 {
-                    percentage = (int)((double)(i + 1) / request.Resources.Count * 100);
+                    percentage = (float)(i + 1) / request.Resources.Count * 100f;
 
                     var resource = request.Resources[i];
                     if (resource.NameDefined)
